refactor(src/utils): move ipfs methods in cache to ipfs utils and small refactor
